% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPreScreenPredictors.R
\name{autoPreScreenPredictors}
\alias{autoPreScreenPredictors}
\title{Pre-Screening for Top Predictors}
\usage{
autoPreScreenPredictors(X, y, k = 10)
}
\arguments{
\item{X}{A data frame containing the predictor variables.}

\item{y}{A vector representing the response variable. It can be continuous or categorical.}

\item{k}{An integer specifying the number of top predictors to return. Default is 10.}
}
\value{
A data frame containing:
\itemize{
  \item \code{predictor}: The name of the predictor variable.
  \item \code{p_value}: The p-value indicating the strength of the relationship between the predictor and the response.
  \item \code{method}: The statistical test used to evaluate the relationship.
}
}
\description{
This function pre-screens a set of predictors to find the top `k` most informative ones in terms of their relationship with a given response variable. It applies different statistical methods depending on the types of predictors and the response.
}
\details{
The function applies the following statistical tests based on the type of predictors and response:
\itemize{
  \item If both the predictor and response are continuous, it uses Pearson Correlation.
  \item If the predictor is binary and the response is continuous, or vice versa, it uses the T-test.
  \item If the predictor is categorical with more than two levels and the response is continuous, it uses ANOVA.
  \item If both the predictor and response are categorical, it uses the Chi-squared test.
}

Before applying these tests, the function checks for and omits rows with missing values, issuing a warning if null values are found.
If `k` is greater than the number of predictors in `X`, it adjusts `k` to the number of columns in `X`, issuing a warning.
}
\examples{
# Create a data frame with different types of predictors and a continuous response
set.seed(123)
X <- data.frame(
  Continuous1 = rnorm(100, 50, 10),
  Continuous2 = rnorm(100, 100, 15),
  Binary1 = sample(c("GroupA", "GroupB"), 100, replace = TRUE),
  Binary2 = rbinom(100, 1, 0.5),
  Categorical = sample(c("Red", "Green", "Blue"), 100, replace = TRUE)
)
y <- rnorm(100, 75, 20)  # continuous response

# Get top 3 predictors for a continuous response
pre_screen_predictors(X, y, k = 3)

# Create a binary response and get top 3 predictors
y_binary <- sample(c("Yes", "No"), 100, replace = TRUE)  # binary response
pre_screen_predictors(X, y_binary, k = 3)

}
\seealso{
\code{cor.test}, \code{t.test}, \code{aov}, \code{chisq.test}
}
