% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensembleModel}
\alias{ensembleModel}
\title{Ensemble Model for Multiple Predictive Models}
\usage{
ensembleModel(y, X, modelTypes, combine = "average")
}
\arguments{
\item{y}{A vector representing the response variable, which can be binary or continuous.}

\item{X}{A matrix of predictors.}

\item{modelTypes}{A character vector specifying the types of models to fit.
Examples include "autoRidgeRegression", "autoLassoRegression", "Elastic_net", etc.}

\item{combine}{A character string indicating the method to combine the model predictions.
Current supported method is "average".}
}
\value{
A list containing the models' predictions and the combined predictions.
}
\description{
This function allows for the ensemble of multiple predictive models, enabling the combination
of their predictions for potentially improved accuracy and robustness. It supports a range of
models defined within the package and can combine their predictions using specified methods.
}
\examples{
\dontrun{
  data <- data.frame(
    response = rnorm(100),
    predictor1 = rnorm(100),
    predictor2 = rnorm(100),
    predictor3 = rnorm(100)
  )
  y <- data$response
  X <- data[, c("predictor1", "predictor2", "predictor3")]
  ensemble_results <- ensembleModel(y = y, X = X,
    modelTypes = c("autoRidgeRegression", "autoLassoRegression", "Elastic_net"),
    combine = "average"
  )
  print(ensemble_results$combined_predictions)
}
}
